Marketplace-Copilot/Marketplace-Copilot/python_services/requirements.txt
psycopg2-binary
requests
beautifulsoup4
scikit-learn
pandas
schedule
lxml
statsmodels
